Commit adds support for the Java module system and changes multiple classes to not use private packages of com.sun. *.
Skins have been changed to use the default skins of
javafx.scene.control.skin.

FluentButtonSkin have an exception of ClassCastException with the
DatePicker.
The SpinnerSample shows an X in the field when some number is selected.

The module-info will have to change if the ControlsFX project changes
the name of their module.